src/requirements.txt
asttokens==2.2.1
backcall==0.2.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
cryptography==39.0.1
cuda-python==12.0.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.9.0
firebase-admin==6.1.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.1
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
more-itertools==9.0.0
msgpack==1.0.4
nest-asyncio==1.5.6
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-pyindex==1.0.9
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
pika==1.3.1
platformdirs==3.0.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
rsa==4.9
six==1.16.0
stack-data==0.6.2
tokenizers==0.13.2
torch==1.13.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
whisper-openai==1.0.0
zipp==3.13.0
